requirements-prod.txt
fastapi==0.111.0
uvicorn==0.29.0
GitPython==3.1.43
loguru==0.7.2
